% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_mean_sd.R
\name{base_mean_sd}
\alias{base_mean_sd}
\title{Mean and SD of Basis Distributions}
\usage{
base_mean_sd(model, base, dim = 1)
}
\arguments{
\item{model}{a fitted GPT model. See \code{\link{fit.gpt}}}

\item{base}{optional: a vector with indices for component distributions}

\item{dim}{dimension of the continuous variable}
}
\description{
Computes summary statistics for basis distributions, conditional on the 
parameter estimates.
}
