% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_mean_sd.R
\name{component_mean_sd}
\alias{component_mean_sd}
\title{Mean and SD of Basis Distributions}
\usage{
component_mean_sd(gpt_fit, component, dim = 1)
}
\arguments{
\item{gpt_fit}{a fitted GPT model. See \code{\link{gpt_fit}}}

\item{component}{optional: a vector with indices for component distributions}

\item{dim}{dimension of the continuous variable}
}
\description{
Computes summary statistics for basis distributions, conditional on the 
parameter estimates.
}
