% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gpt.R
\name{gen.gpt}
\alias{gen.gpt}
\title{Generate Single Data Set}
\usage{
gen.gpt(n, theta, eta, file, latent, restrictions = NULL)
}
\arguments{
\item{n}{vector of the length of number of trees with n per tree}

\item{theta}{vector of true MPT parameter values (consider also restricted parameters!). Values will be appropriatly reordered if a named vector ist given (otherwise, check order!).}

\item{eta}{vector of true continuous parameters. Values will be appropriatly reordered if a named vector ist given (otherwise, check order!).}

\item{file}{a character vector specifying the path to the model file}

\item{latent}{type of latent continuous distribution (one of \code{"normal"}, \code{"exgauss"}, \code{"exwald"}, \code{"gamma"}, \code{"weibull"}, \code{"lognormal"}, \code{"wald"}, or \code{beta}). Can be a  vector if multiple continuous variables \code{y} have different distributions (e.g., \code{latent = c("normal", "gamma")})}

\item{restrictions}{list with parameter restrictions (e.g., \code{list("g=0.5", "mean1=mean2=100")})}
}
\description{
Generate Single Data Set
}
\examples{
###### 2-High-Threshold Model (+2 normal distributions) ######

### parameters
n <- c(targets=20, lures=20)     # number of items
theta <- c(do=.6, g=.5)          # MPT parameters
eta <- c(m1_d=100, m1_g=110, s1=10,
         m2_d=30, m2_g=50, s2=5)     # normal distributions
file <- paste0(path.package("gpt"), "/models/2htm_2normal.txt")

gen <- gen.gpt(n=n, theta=theta, eta=eta, file=file,
              latent=c("normal", "normal"), restrictions=list("do=dn"))
head(gen) 
# check means of latent continuous distributions:
tapply(gen$y.1, gen$state, mean)

}
\seealso{
gen.gpt.sample
}
