% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{hist.gpt.fit}
\alias{hist.gpt.fit}
\title{Plot Predicted and Observed Densities}
\usage{
\method{hist}{gpt.fit}(x, dim = 1, freq = "tree", group = 1, ...)
}
\arguments{
\item{x}{fitted model return from \code{\link{fit.gpt}}}

\item{dim}{only for multivariate continuous data: dimension for prediction}

\item{freq}{how to normalize histogram and density. Either per category (\code{freq = "cat"}), per tree (\code{freq = "tree"}), or in absolute frequencies (\code{freq = "freq"})}

\item{group}{select group by an index, e.g., \code{group=1} (if mutliple groups were fitted)}

\item{...}{further arguments passt to \code{\link{hist}}}

\item{which}{group to plot (not tested!)}
}
\description{
Plot predicted against observed continuous distributions
}
