% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_test.R
\name{lr.test}
\alias{lr.test}
\title{Likelihood-Ratio Test}
\usage{
lr.test(model, nested, EM = TRUE)
}
\arguments{
\item{model}{the more general model (see \code{\link{fit.gpt}})}

\item{nested}{the nested model (i.e., a model that differs only in additional order constraints). Note that the nestedness of models is \italic{not} checked internally}

\item{EM}{whether to use the expectation-maximization fit or that of the gradient-based optimization}
}
\description{
Test whether the increase of misfit of a nested model is significant.
}
\seealso{
\code{\link{gpt.select}}
}
