% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_goodness_of_fit.R
\name{model_fit_discrete}
\alias{model_fit_discrete}
\title{Goodness of Fit of GPT Models}
\usage{
model_fit_discrete(model, intervals = 10, breaks)
}
\arguments{
\item{model}{a fitted gpt model (see \code{\link{fit.gpt}})}

\item{intervals}{number of intervals used for testing}

\item{breaks}{vector with (increasing) boundary values of intervals used for grouping}
}
\description{
Uses a fixed numbers of histogram bins to compare GPT predictions against the empirical distribution (only for one continuous variable and group=1).
}
\seealso{
\code{\link{lr.test}}
}
