% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.gpt.fit}
\alias{print.gpt.fit}
\title{Print GPT Model}
\usage{
\method{print}{gpt.fit}(x, ci = 0.95, digits = 3, ...)
}
\arguments{
\item{x}{fitted model of \code{\link{fit.gpt}}}

\item{ci}{probability for confidence interval}

\item{digits}{number of digits to be printed}

\item{...}{ignored}
}
\description{
Either estimated parameters (single data set) or summary of estimates (if model was fitted for a sample of participants using \code{group})
}
