% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_gpt_fit.R
\name{print.gpt_fit}
\alias{print.gpt_fit}
\title{Print GPT Model}
\usage{
\method{print}{gpt_fit}(x, ci = 0.95, digits = 3, EM = FALSE, ...)
}
\arguments{
\item{x}{fitted model of \code{\link{gpt_fit}}}

\item{ci}{probability for confidence interval}

\item{digits}{number of digits to be printed}

\item{EM}{whether to use parameter estimates and likelihood from
expectation-maximization (EM) fit or from the gradient-based optimization}

\item{...}{ignored}
}
\description{
Either estimated parameters (single data set) or summary of estimates (if model was fitted for a sample of participants using \code{group})
}
