% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale bounded variable to given range}
\usage{
rescale(y, min = 0, max = 1)
}
\arguments{
\item{y}{vector of numeric values}

\item{min}{numeric value defining the minimum of the target range}

\item{max}{numeric value defining the maximum of target range}
}
\description{
Uses a linear transformation to rescale variable to a specific range (e.g., [0,1] to adhere to a beta distribution). The new minimum/maximum will be equal to the lower/upper bound.
}

