% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_gpt.R
\name{select_gpt}
\alias{select_gpt}
\title{Model Selection for GPT Models}
\usage{
select_gpt(..., gpt_fits = list())
}
\arguments{
\item{...}{fitted GPT models (see \code{\link{gpt_fit}})}

\item{gpt_fits}{a list of fitted GPT models that are added to \code{...}}
}
\description{
Currently, works only for models fitted to a single dataset
}
\seealso{
\code{\link{test_nested}}
}
