% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_gpt_fit.R
\name{summary.gpt_fit}
\alias{summary.gpt_fit}
\title{Summary for Fitted GPT Models}
\usage{
\method{summary}{gpt_fit}(object, EM = FALSE, ...)
}
\arguments{
\item{object}{a fitted GPT model (see \code{\link{gpt_fit}})}

\item{EM}{whether to use parameter estimates and likelihood from
expectation-maximization (EM) fit or from the gradient-based optimization}

\item{...}{further arguments passed to \code{\link{test_fit}}}
}
\description{
Shows parameter estimates and goodness of fit.
}
