% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_nested_LR.R
\name{test_nested}
\alias{test_nested}
\title{Likelihood-Ratio Test}
\usage{
test_nested(gpt_fit, nested, EM = TRUE)
}
\arguments{
\item{gpt_fit}{the more general GPT model (see \code{\link{gpt_fit}})}

\item{nested}{a nested GPT model (i.e., a model that differs only in additional order constraints). 
Note that the nestedness of models is not checked internally.}

\item{EM}{whether to use parameter estimates and likelihood from
expectation-maximization (EM) fit or from the gradient-based optimization}
}
\description{
Test whether the increase of misfit of a nested model is significant.
}
\seealso{
\code{\link{select_gpt}}, \code{\link{test_fit}}
}
