% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_gpt.R
\name{update.gpt_fit}
\alias{update.gpt_fit}
\title{Re-Fit GPT Model with Additional Constraints}
\usage{
\method{update}{gpt_fit}(object, restrictions, ...)
}
\arguments{
\item{object}{a fitted GPT model (see \code{\link{gpt_fit}})}

\item{restrictions}{a list of additional parameter constraints that are added 
to the original model}

\item{...}{further arguments passed to \code{\link{gpt_fit}} for fitting the nested GPT model}
}
\description{
Fits nested versions of GPT models.
}
\seealso{
\code{\link{gpt_fit}}
}
