% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cdf.R
\name{cdf}
\alias{cdf}
\title{Plot Predicted and Empirical Cumulative Densities}
\usage{
cdf(model, ks.test = TRUE, dim = 1, group = 1, ...)
}
\arguments{
\item{model}{fitted GPT model (\link{fit.gpt})}

\item{ks.test}{whether to perform Kolmogorov-Smirnov Tests (\link{ks.test})}

\item{dim}{only for multivariate continuous data: dimension for prediction}

\item{group}{select group by an index, e.g., \code{group=1} (if mutliple groups were fitted)}

\item{...}{further arguments passt to \code{\link{plot}}}
}
\description{
Plot predicted against observed cumulative densities (separately per observed category)
}
