% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{gpt.select}
\alias{gpt.select}
\title{Compare GPT Models}
\usage{
gpt.select(model.list)
}
\arguments{
\item{model.list}{list of fitted gpt-models (see \code{\link{fit.gpt}})}
}
\description{
Currently, only functional for models fitted to single datasets
}
