% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{hist.gpt.fit}
\alias{hist.gpt.fit}
\title{Plot Predicted and Observed Densities}
\usage{
\method{hist}{gpt.fit}(x, dim = 1, ...)
}
\arguments{
\item{x}{fitted model return from \code{\link{fit.gpt}}}

\item{dim}{only for multivariate continuous data: dimension for prediction}

\item{...}{further arguments passt to \code{\link{hist}}}
}
\description{
Plot predicted against observed continuous distributions
}

